project/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.4
Django
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
drf-yasg
flake8==7.0.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
josepy==1.14.0
kombu==5.3.5
lockfile==0.12.2
MarkupSafe==2.1.5
mccabe==0.7.0
mozilla-django-oidc==3.0.0
olefile==0.46
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-decouple==3.8
pytz==2024.1
celery[redis]==5.3.6
redis==5.0.1
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
sentry-sdk==2.8.0
channels==4.1.0
uvicorn==0.27.0
websockets==12.0
channels_redis==4.2.0
jpype1~=1.5.2
opencv-python==4.12.0.88
numpy==2.2.6